Introduced subscription handles.
Summary:
- a non-RAII (read: minimal memory overhead) subscription handles,
- I didn't want to stuff the very same client pointer into every handle..., what it client goes out of the scope before the handle?
- I don't want server-side to be forced to store 1M copies of client ptr as well,

Things to discuss:
- shall we make them noncopyable now, before it's too late? if one can guarantee noncopyability and that Acknowledge always happens-before Unsubscribe, then we can use unique_ptr in place of shared one

Test Plan: `make -j all check`

Reviewers: rocketspeed, pja

Reviewed By: pja

Subscribers: dhruba, eduardosuarez, pja, rocketspeed

Differential Revision: https://reviews.facebook.net/D38499

+++ src/messages/messages.h
@@ -642,7 +642,7 @@ inline bool ValidateEnum(MessageGoodbye::OriginType e) {
  * ID of the subscription the message belongs to. IDs are uniquely identify
  * subscription within a stream.
  */
-typedef uint32_t SubscriptionID;
+typedef uint64_t SubscriptionID;
 
 /** A request to subscribe to provided topic with given parameters. */
 class MessageSubscribe final : public Message {
